requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
#redis==5.0.1  # https://github.com/redis/redis-py
#hiredis==2.2.3  # https://github.com/redis/hiredis-py
whitenoise==6.6.0  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
django==4.2.7  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth>=0.55.0  # https://github.com/pennersr/django-allauth

# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.3.1  # https://github.com/adamchainz/django-cors-headers
dj-rest-auth[with_social]==5.0.1  # https://github.com/iMerica/dj-rest-auth
djangorestframework-simplejwt==5.3.0

# DRF-spectacular for api documentation
drf-spectacular>=0.26.5  # https://github.com/tfranzel/drf-spectacular

django-taggit==5.0.1
django-filter==23.3  # https://github.com/carltongibson/django-filter
